Move used types from data to api.d.ts
