+++ src/logic/api.d.ts
@@ -1,5 +1,3 @@
-import {PositionAnswer, Vote} from './data';
-
 export interface ApiSearchPolitician {
   id: number;
   label: string;
@@ -245,3 +243,9 @@ export type TopicIcon = {
   label: string;
   icon: string;
 };
+
+export type Vote = 'yes' | 'no' | 'abstain' | 'no_show';
+
+export type PollResult = 'yes' | 'no';
+
+export type PositionAnswer = 'agree' | 'disagree' | 'neutral';
